River expiry application
* The river expiry application is in the maintenance category of the
SwiftRiver-Core applications. It is designed to be run as a cron job.
It checks for rivers that have expired or are about to expire and sends
email notifications to its owners. It uses the sendmail program for
sending email. The config file therefore needs to specify the location
of the sendmail binary. Closes #12