Confirm drops only after they are published
Issue with media extractor where if the drop publisher thread hangs due
to an exception, incoming drops are still confirmed and consequently
lost since they are never published back an accumulate in the in memory
queue.

Change made to Publisher to fire a callback if available once a drop is
successfully published.